requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn==1.6.1
xgboost>=1.7
joblib
dnspython
python-whois
shap
matplotlib

